split into internal and structured
