+++ crashtracker/src/crash_info/internal/metadata.rs
