+++ crashtracker/src/crash_info/internal/stacktrace.rs
